order endpoints added

+++ apps/orders/views.py
@@ -6,15 +6,13 @@ from rest_framework.viewsets import (
     ModelViewSet,
     ViewSet,
 )
-from rest_framework.mixins import (
-    ListModelMixin,
-)
+from rest_framework.views import APIView
 from rest_framework.generics import (
-    GenericAPIView,
     ListAPIView,
 )
 from django.shortcuts import get_object_or_404
-from django.db.models import QuerySet, Count
+from django.db.models import QuerySet, Count, Sum, F
+from django.db import transaction
 from rest_framework.permissions import (
     IsAuthenticatedOrReadOnly,
     IsAdminUser,
@@ -40,14 +38,16 @@ from .serializers import (
     CartItemCreateSerializer,
     CartItemUpdateSerializer,
     CustomUserCartSerializer,
-    OrderListSerializer,
+    OrderListCreateSerializer,
 )
 from .permissions import IsOwnerOrReadOnly
 from apps.users.models import CustomUser
 
 
 class ReviewViewSet(ModelViewSet):
-    """ViewSet for Review model."""
+    """
+    Handle all types of requests to reviews.
+    """
     serializer_class = ReviewSerializer
     permission_classes = [IsAuthenticatedOrReadOnly, IsOwnerOrReadOnly]
 
@@ -88,8 +88,8 @@ class CartItemViewSet(ViewSet):
         """
         if self.action == "list":
             permission_classes = [IsAdminUser]
-        elif self.action == "retrieve":
-            permission_classes = [IsAdminUser, IsOwnerOrReadOnly]
+        elif self.action in ["retrieve", "partial_update", "destroy"]:
+            permission_classes = [IsOwnerOrReadOnly]
         elif self.action == "create":
             permission_classes = [IsAuthenticated]
         else:
@@ -104,18 +104,29 @@ class CartItemViewSet(ViewSet):
     ) -> DRFResponse:
         """
         Handles GET requests to list of cart items.
-        Returns carts of all users.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing a list of all users cart items.
         """
 
-        # if not request.user.is_staff:
-        #     raise PermissionDenied(
-        #         "You can't access cart items of other users."
-        #     )
+        if not request.user.is_staff:
+            raise PermissionDenied(
+                "You can't access cart items of other users."
+            )
 
         users: QuerySet[CustomUser] = CustomUser.objects.prefetch_related(
             "cart_items").annotate(
                 total_positions=Count("cart_items__id")
-            )
+        )
 
         paginator: PageNumberPagination = self.pagination_class()
         page = paginator.paginate_queryset(users, request=request)
@@ -125,7 +136,7 @@ class CartItemViewSet(ViewSet):
             many=True,
         )
 
-        return DRFResponse(data=serializer.data, status=HTTP_200_OK)
+        return paginator.get_paginated_response(serializer.data)
 
     def retrieve(
         self,
@@ -135,17 +146,31 @@ class CartItemViewSet(ViewSet):
         **kwargs: dict[Any, Any],
     ) -> DRFResponse:
         """
-        Handles GET requests to list of cart items.
-        Returns a list of cart items of a user by specified id.
+        Handles GET requests to cart items of a single user.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            pk: int
+                User's id.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing list of specified user's cart items.
         """
+
         user: CustomUser = get_object_or_404(CustomUser, pk=pk)
 
         # Check if the request was sent by staff or cart's owner:
 
-        # if request.user != user and not request.user.is_staff:
-        #     raise PermissionDenied(
-        #         "You can't access cart items of other users."
-        #     )
+        if request.user != user and not request.user.is_staff:
+            raise PermissionDenied(
+                "You can't access cart items of other users."
+            )
 
         cart_items: QuerySet[CartItem] = CartItem.objects.filter(
             user=user,
@@ -171,6 +196,18 @@ class CartItemViewSet(ViewSet):
     ) -> DRFResponse:
         """
         Handles POST requests to add a new item to user's cart.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing information about added cart item.
         """
 
         request_data = request.data
@@ -222,7 +259,22 @@ class CartItemViewSet(ViewSet):
         **kwargs: dict[Any, Any],
     ) -> DRFResponse:
         """
-        Handles PATCH requests to update an existing CartItem instance.
+        Handles PATCH requests to partially update info
+        about existing item in a cart.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            pk: int
+                Cart item id.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing info about an updated item.
         """
         try:
             existing_cartitem: CartItem = CartItem.objects.get(pk=pk)
@@ -233,6 +285,8 @@ class CartItemViewSet(ViewSet):
                 },
                 status=HTTP_404_NOT_FOUND,
             )
+
+        self.check_object_permissions(request=request, obj=existing_cartitem)
 
         serializer: CartItemUpdateSerializer = CartItemUpdateSerializer(
             instance=existing_cartitem,
@@ -255,7 +309,21 @@ class CartItemViewSet(ViewSet):
         **kwargs: dict[Any, Any],
     ) -> DRFResponse:
         """
-        Handles DELETE requests to destroy an existing CartItem instance.
+        Handles DELETE requests to cart items.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            pk: int
+                Cart item id.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                Status of the response.
         """
         try:
             existing_cartitem: CartItem = CartItem.objects.get(pk=pk)
@@ -266,18 +334,19 @@ class CartItemViewSet(ViewSet):
                 },
                 status=HTTP_404_NOT_FOUND,
             )
+        self.check_object_permissions(request=request, obj=existing_cartitem)
         existing_cartitem.delete()
         return DRFResponse(
             status=HTTP_204_NO_CONTENT,
         )
 
 
-class OrderViewSet(ListAPIView):
+class OrderListView(ListAPIView):
     """
     Handles GET requests to Order model.
     """
 
-    serializer_class = OrderListSerializer
+    serializer_class = OrderListCreateSerializer
     permission_classes = [IsAuthenticated]
 
     def get_queryset(self) -> QuerySet[Order]:
@@ -286,8 +355,11 @@ class OrderViewSet(ListAPIView):
 
         user_orders = Order.objects.filter(user=user).prefetch_related(
             "order_items").annotate(
-                total_positions=Count("order_items__id")
-            )
+                total_positions=Count("order_items__id"),
+                total_price=Sum(
+                    F("order_items__price") * F("order_items__quantity")
+                ),
+        )
 
         return user_orders
 
@@ -297,9 +369,128 @@ class OrderViewSet(ListAPIView):
         *args: tuple[Any, ...],
         **kwargs: dict[Any, Any],
     ) -> DRFResponse:
-        """Handles GET requests to a list of specified user's orders."""
+        """
+        Handles GET requests to a list of specified user's orders.
 
-        if request.user.id != self.kwargs.get("user_id") and not request.user.is_staff:
+        Parameters:
+            request: DRFRequest
+                The request object.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing a list of user's orders.
+        """
+
+        if (
+            request.user.id != self.kwargs.get("user_id") and
+            not request.user.is_staff
+        ):
             raise PermissionDenied("You can't access orders of other users.")
 
         return super().get(request, *args, **kwargs)
+
+
+class OrderCreateView(APIView):
+    """
+    View to create a new order from existing cart items.
+    """
+
+    permission_classes = [IsAuthenticated]
+
+    def post(
+        self,
+        request: DRFRequest,
+        *args: tuple[Any, ...],
+        **kwargs: dict[Any, Any],
+    ) -> DRFResponse:
+        """
+        Handles POST requests to create a new order from exisiting items.
+        Order is done by the user sending the request.
+
+        Parameters:
+            request: DRFRequest
+                The request object.
+            *args: list
+                Additional positional arguments.
+            **kwargs: dict
+                Additional keyword arguments.
+
+        Returns:
+            DRFResponse
+                A response containing info about a new order.
+        """
+
+        with transaction.atomic():
+            user = request.user
+            cart_items = CartItem.objects.filter(
+                user=user).select_related("product")
+
+            if not cart_items.exists():
+                return DRFResponse(
+                    data={
+                        "cart_items": ["Your cart is empty."],
+                    },
+                    status=HTTP_400_BAD_REQUEST,
+                )
+
+            phone_number: str = request.data.get("phone_number")
+            delivery_address: str = request.data.get("delivery_address")
+            status: str = "P"
+
+            if not phone_number or not delivery_address:
+                return DRFResponse(
+                    data={
+                        "phone_number": ["Phone number can't be null."],
+                        "delivery_address": ["Delivery address can't be null."]
+                    },
+                    status=HTTP_404_NOT_FOUND,
+                )
+
+            order: Order = Order.objects.create(
+                user=request.user,
+                phone_number=phone_number,
+                delivery_address=delivery_address,
+                status=status,
+            )
+
+            order_items: list[OrderItem] = []
+            total_price: float = 0
+            total_positions: int = 0
+
+            for item in cart_items:
+                product: Product = item.product
+                name: str = item.product.name
+                price: float = item.product.price
+                quantity: int = item.quantity
+                total_price += round(price * quantity, 2)
+                total_positions += quantity
+
+                order_items.append(
+                    OrderItem(
+                        order=order,
+                        product=product,
+                        name=name,
+                        price=price,
+                        quantity=quantity,
+                    )
+                )
+
+            OrderItem.objects.bulk_create(order_items)
+            cart_items.delete()
+
+            serializer: OrderListCreateSerializer = OrderListCreateSerializer(
+                instance=order,
+                context={
+                    "total_price": total_price,
+                    "total_positions": total_positions,
+                    },
+            )
+
+            return DRFResponse(
+                data=serializer.data,
+                status=HTTP_201_CREATED,
+            )
